refactor: state render-> movie state로 변경
